[Refactor] #172 - 바텀시트 컴포넌트 분리
